requirements.txt
appdirs==1.4.4
attrs==19.3.0
cfgv==3.1.0
colorama==0.4.3
distlib==0.3.1
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
pymysql==1.1.0
identify==1.4.20
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.4.0
pathspec==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.2
python-dotenv==0.13.0
PyYAML==5.4
regex==2020.6.8
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.25
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
elasticsearch==8.9.0
requests==2.31.0